DRT: extract RequestInsertionScheduler interface
